+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Add option to multiply width by device pixel ratio
 - Provide new set of default widths
 - Add version number information as comment to js
+- Remove code to replace divs with imgs - act on imgs directly
+- Add option to specify a target attribute instead of src
+- Add option to specify a source attribute instead of data-src
 
 ### 0.5.0 (2015/01/09 19:30 +00:00)
 - [#140](https://github.com/BBC-News/Imager.js/pull/140) Respect alt attribute of placeholder img (@wildlyinaccurate)